First draft using the DRV8262 H-Bridge
This design is complete and ready for manufacture, however it has
become clear that the DRV8262 over current protection is not compatible
with capacitive loads because it works by engaging both high-side
FETs which shorts the outputs.  That's fine for braking an inductive
load like a motor but it will rapidly discharge a capacitive load
instead which is the opposite of what we're trying to do.

Don't build this version.