Revised layout using LM5108 and discrete FETs
LM5108 is less expensive but it requires more board area.  In lieu of
active current limiting, we rely on the LM5158 hiccup overload
protection.  The FETs are intentionally oversized to handle large
current pulses when switching the polarity of the capacitive load.

This design is complete and ready for manufacture.